update index.js and finish lesson 1-2
